Splits code into modules (homework 5.14. Modulation)
